requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.4
langchain-huggingface>=0.1.0
langgraph>=0.2.50
chromadb>=0.5.0
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.4.0
pandas>=2.2.0
pyarrow>=15.0.0
datasets>=2.20.0
huggingface_hub>=0.24.0
pydantic>=2.7.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0
